@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@babel/core": "^7.8.7",
+    "@babel/helper-define-map": "^7.16.7",
     "@babel/helper-regex": "^7.10.5",
     "@babel/plugin-transform-typescript": "^7.8.7",
     "@emotion/babel-preset-css-prop": "^10.0.27",
@@ -86,14 +87,14 @@
   "devDependencies": {
     "@babel/preset-env": "^7.15.6",
     "@svgr/webpack": "5.4.0",
-    "@typescript-eslint/eslint-plugin": "^2.23.0",
-    "@typescript-eslint/parser": "^2.23.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "eslint": "^8.7.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "husky": "4.2.4",
     "lint-staged": "10.1.3",
     "next-sitemap": "^1.9.12"
