@@ -1,5 +1,6 @@
 declare module "*.svg" {
-  import React = require("react"); // eslint-disable-line no-unused-vars
+  import React = require("react");
+
   export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>;
   const src: string;
   export default src;
