@@ -105,7 +105,10 @@ function HeroCard({ topic, description, button, links }: Props) {
         links.map((link, index) => (
           // eslint-disable-next-line react/no-array-index-key
           <StyledLink key={index}>
-            <StyledAnchor href={link.href} target={link.external && "_blank"}>
+            <StyledAnchor
+              href={link.href}
+              target={link.external ? "_blank" : undefined}
+            >
               {link.text} &gt;
             </StyledAnchor>
           </StyledLink>
